refactor : 구글맵 page별로 스타일 변경할 수 있도록 로직 수정
